# AI客服/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ruitalk Seller Terminal API",
    "version": "1.0.0",
    "description": "AI-powered multilingual customer service system API",
    "contact": {
      "name": "Ruitalk Team",
      "email": "team@ruitalk.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Development"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Alternate"
    }
  ],
  "paths": {
...
    "/api/v1/platforms": {
      "get": {
        "summary": "获取平台配置状态",
        "description": "获取平台配置状态",
        "operationId": "get__api_v1_platforms",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/sync": {
      "post": {
        "summary": "手动触发平台同步",
        "description": "手动触发平台同步",
        "operationId": "post__api_v1_sync",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/sync/status": {
      "get": {
        "summary": "获取同步状态",
        "description": "获取同步状态",
        "operationId": "get__api_v1_sync_status",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "summary": "获取订单列表（优先读本地同步缓存）",
        "description": "获取订单列表（优先读本地同步缓存）",
        "operationId": "get__api_v1_orders",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "summary": "获取订单详情",
        "description": "获取订单详情",
        "operationId": "get__api_v1_orders_order_id",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: order_id"
          }
        ]
      }
    },
    "/api/v1/returns": {
      "get": {
        "summary": "获取退换货列表",
        "description": "获取退换货列表",
        "operationId": "get__api_v1_returns",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "创建售后单（写入本地库 + 调用外部API）",
        "description": "创建售后单（写入本地库 + 调用外部API）",
        "operationId": "post__api_v1_returns",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews": {
      "get": {
        "summary": "获取评价列表",
        "description": "获取评价列表",
        "operationId": "get__api_v1_reviews",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/reviews/reply": {
      "post": {
        "summary": "回复评价",
        "description": "回复评价",
        "operationId": "post__api_v1_reviews_reply",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/logistics/{tracking_number}": {
      "get": {
        "summary": "查询物流轨迹",
        "description": "查询物流轨迹",
        "operationId": "get__api_v1_logistics_tracking_number",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "name": "tracking_number",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "路径参数: tracking_number"
          }
        ]
      }
    },
    "/api/v1/dashboard/stats": {
      "get": {
        "summary": "获取仪表盘统计数据",
        "description": "获取仪表盘统计数据",
        "operationId": "get__api_v1_dashboard_stats",
        "tags": [
          "Platform"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
...
  }
}